# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Backend Service API"
  description: "API documentation for Customers, Mechanics, Service Tickets, and Inventory"
  version: "1.0.0"

# -------------------------
# Server configuration
# -------------------------
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# API Paths (Endpoints)
# =========================
paths:
...
  /service_tickets/:
    post:
      summary: Create a new service ticket
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ServiceTicketInput'
      responses:
        201:
          description: Service ticket created
          schema:
            $ref: '#/definitions/ServiceTicket'
    get:
      summary: Get all service tickets
      responses:
        200:
          description: List of service tickets
          schema:
            type: array
            items:
              $ref: '#/definitions/ServiceTicket'
  /service_tickets/{ticket_id}:
    get:
      summary: Get a single service ticket
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
      responses:
        200:
          description: Service ticket object
          schema:
            $ref: '#/definitions/ServiceTicket'
    patch:
      summary: Update service ticket fields
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ServiceTicketInput'
      responses:
        200:
          description: Updated service ticket
          schema:
            $ref: '#/definitions/ServiceTicket'
    delete:
      summary: Delete a service ticket
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
      responses:
        200:
          description: Success message
          schema:
            $ref: '#/definitions/Message'
...
definitions:
...
  ServiceTicketInput:
    type: object
    properties:
      customer_id: { type: integer } # ID of the customer
      description: { type: string }  # Description of the service

  # Used for returning service ticket data (output)
  ServiceTicket:
    type: object
    properties:
      id: { type: integer }          # Unique ticket ID
      customer_id: { type: integer }
      description: { type: string }

  # Used for adding a part to a service ticket (input)
...
  Message:
    type: object
    properties:
      message: { type: string }          # Message text